Create README.md
The truth is I'm working my butt off to learn, eventually, good enough skills to get paid to work in cybersecurity. I know everyone just rolled their eyes, but I have a condition where the bones in my back are deteriorating and eventually I will be in a wheelchair. I have made it through nursing school (RN), quit and became a dump truck driver to an operator to superintindant in just over a year with no experience; not saying this is going to be easy. I'm just saying I'm extremely tenacious and I definitly can out work anyone even from a wheelchair. I look forward to working with some of you on something someday.